style: 불필요 코드 제거, 엔티티 수정 - 생성일, 수정일 코드 수정 - project - progress 제거 - member_project - status 추가

+++ src/main/java/com/douzone/prosync/task/dto/request/TaskPatchDto.java
@@ -38,7 +38,7 @@ public class TaskPatchDto {
     private String endDate;
 
     @ApiModelProperty(value = "업무상태식별자", example = "1")
-    private Integer taskStatusId;
+    private Long taskStatusId;
 
     @ApiModelProperty(hidden = true)
     private LocalDateTime modifiedAt;
