+++ src/main/java/com/douzone/prosync/task/dto/request/TaskPostDto.java
@@ -42,7 +42,7 @@ public class TaskPostDto {
         private String endDate;
 
         @ApiModelProperty(value = "업무상태", example = "1")
-        private Integer taskStatusId;
+        private Long taskStatusId;
 
         @ApiModelProperty(hidden = true)
         private LocalDateTime createdAt;
@@ -51,7 +51,7 @@ public class TaskPostDto {
                 this.createdAt = createdAt;
         }
 
-        public void setTaskStatusId(Integer taskStatusId) {
+        public void setTaskStatusId(Long taskStatusId) {
                 this.taskStatusId = taskStatusId;
         }
 }
